BCOP-2976 fix logging

+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/ApplicationServices/AffidavitPostProcessingService.cs
@@ -11,10 +11,13 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
+using System.Web.Mvc;
 using EntityFrameworkMapping.Broadcast;
+using Services.Broadcast.Repositories;
 using Tam.Maestro.Common;
 using Tam.Maestro.Services.Cable.Entities;
 using Tam.Maestro.Services.Cable.SystemComponentParameters;
+using Tam.Maestro.Services.ContractInterfaces;
 
 namespace Services.Broadcast.ApplicationServices
 {
@@ -23,7 +26,7 @@ namespace Services.Broadcast.ApplicationServices
         /// <summary>
         /// Downloads the WWTV processed files and calls the affidavit processing service
         /// </summary>
-        void DownloadAndProcessWWTVFiles(string userName, out List<string> filesFailedDownload);
+        void DownloadAndProcessWWTVFiles(string userName);
 
         /// <summary>
         /// Process and WWTV post processing file
@@ -37,9 +40,9 @@ namespace Services.Broadcast.ApplicationServices
     public class AffidavitPostProcessingService : IAffidavitPostProcessingService
     {
         private readonly IBroadcastAudiencesCache _AudienceCache;
-        private readonly IAffidavitValidationEngine _AffidavitValidationEngine;
         private readonly IAffidavitEmailSenderService _AffidavitEmailSenderService;
         private readonly IAffidavitService _AffidavidService;
+        private readonly IAffidavitRepository _AffidavitRepository;
 
         private const string VALID_INCOMING_FILE_EXTENSION = ".txt";
         private const string HTTP_ACCEPT_HEADER = "application/json";
@@ -50,14 +53,13 @@ namespace Services.Broadcast.ApplicationServices
         public AffidavitPostProcessingService(
             IBroadcastAudiencesCache audienceCache,
             IDataRepositoryFactory broadcastDataRepositoryFactory,
-            IAffidavitValidationEngine affidavitValidationEngine,
             IAffidavitEmailSenderService affidavitEmailSenderService,
             IAffidavitService affidavidService)
         {
             _AudienceCache = audienceCache;
-            _AffidavitValidationEngine = affidavitValidationEngine;
             _AffidavitEmailSenderService = affidavitEmailSenderService;
             _AffidavidService = affidavidService;
+            _AffidavitRepository = broadcastDataRepositoryFactory.GetDataRepository<IAffidavitRepository>();
         }
 
 
@@ -66,12 +68,21 @@ namespace Services.Broadcast.ApplicationServices
         /// Downloads the WWTV processed files and calls the affidavit processing service
         /// return true if download success, false if download fails (use for loggin)
         /// This involves FTP 
-        /// 
         /// </summary>
-        public void DownloadAndProcessWWTVFiles(string userName,out List<string> filesFailedDownload)
+        public void DownloadAndProcessWWTVFiles(string userName)
         {
-            filesFailedDownload = new List<string>();
-            List<string> filesToProcess = _GetWWTVFTPFileNames();
+            var filesFailedDownload = new List<string>();
+            List<string> filesToProcess;
+            try
+            {
+                filesToProcess = _GetWWTVFTPFileNames();
+            }
+            catch (Exception e)
+            {
+                _ProceseTotalFTPFailure(e);
+                throw;
+            }
+
             foreach (var file in filesToProcess)
             {
                 string filePath = $"{Path.GetTempPath()}{file}";
@@ -79,45 +90,94 @@ namespace Services.Broadcast.ApplicationServices
                 {
                     _DownloadFileFromWWTVFtp(file, filePath);
                 }
-                catch 
+                catch (Exception e)
                 {               
                     // cannot download file
-                    filesFailedDownload.Add(file);
+                    filesFailedDownload.Add(string.Format("{0} :: Reason -> {1}",file,e.Message));
                     continue;   // skip to next file 
                 }
                 AffidavitSaveResult response = null;
 
-                AffidavitSaveRequest affidavitFile = ParseWWTVFile(filePath);
-                                
-                if (affidavitFile == null)
+                AffidavitSaveRequest affidavitSaveRequest = ParseWWTVFile(filePath);
+                if (!string.IsNullOrEmpty(_errorMessage))
                 {
                     _ProcessError(filePath);
-                    return;
+                    continue;
                 }
                 try
                 {
-                    response = _AffidavidService.SaveAffidavit(affidavitFile, userName, DateTime.Now);
+                    _AffidavidService.SaveAffidavit(affidavitSaveRequest, userName, DateTime.Now);
                 }
                 catch (Exception e)
                 {
                     _errorMessage = "Error saving affidavit:\n\n" + e.ToString();
                     _ProcessError(filePath);
-                    return;
+                    continue;
                 }
 
-                _DeleteWWTVFTPFile(Path.GetFileName(filePath));
+                try
+                {
+                    _DeleteWWTVFTPFile(Path.GetFileName(filePath));
+                }
+                catch (Exception e)
+                {
+                    _errorMessage = "Error deleting affidavit file from FTP site:\n\n" + e.ToString();
+                    _ProcessError(filePath, false);
+                    continue;
+                }
+            }
+
+            if (filesFailedDownload.Any())
+            {
+                _ProcessFailedFiles(filesFailedDownload);
             }
         }
 
-        private void _ProcessError(string filePath)
+        private void _ProceseTotalFTPFailure(Exception e)
+        {
+            var emailBody =
+                "There was an error reading from or connecting to the FTP server. \n\nHere is some technical information." + e;
+            _AffidavitEmailSenderService.Send(emailBody);
+        }
+        private void _ProcessFailedFiles(List<string> filesFailedDownload)
+        {
+            var emailBody = "The following file(s) could not be downloaded.\n\n";
+            foreach (var file in filesFailedDownload)
+            {
+                emailBody += string.Format("{0}\n",file);
+
+            }
+            _AffidavitEmailSenderService.Send(emailBody);
+        }
+
+        private void _ProcessError(string filePath,bool deleteFtpFile = true)
         {
             var invalidFilePath = _MoveFileToInvalidFilesFolder(filePath);
 
             var emailBody = _CreateInvalidFileEmailBody(_errorMessage, invalidFilePath);
 
-            _AffidavitEmailSenderService.Send(emailBody);
+            _AffidavitEmailSenderService.Send(emailBody, "WWTV File Failed");
 
-            _DeleteWWTVFTPFile(Path.GetFileName(filePath));
+            if (deleteFtpFile)
+                _DeleteWWTVFTPFile(Path.GetFileName(filePath));
+
+            _LogAffidavitError(filePath,_errorMessage);
+        }
+
+        private void _LogAffidavitError(string filePath,string errorMessage)
+        {
+            var affidavitFile = new AffidavitFile();
+            affidavitFile.FileName = Path.GetFileName(filePath);
+            affidavitFile.Status = AffidaviteFileProcessingStatus.Invalid;
+            affidavitFile.FileHash = HashGenerator.ComputeHash(filePath.ToByteArray()); // just so there is something
+            affidavitFile.CreatedDate = DateTime.Now;
+            affidavitFile.SourceId = (int)AffidaviteFileSource.Strata;
+
+            var problem = new AffidavitFileProblem();
+            problem.ProblemDescription = errorMessage;
+
+            affidavitFile.AffidavitFileProblems.Add(problem);
+            var id = _AffidavitRepository.SaveAffidavitFile(affidavitFile);
         }
 
         /// <summary>
@@ -127,45 +187,47 @@ namespace Services.Broadcast.ApplicationServices
         /// <returns>BaseResponse object containing the new id of the affidavit_file</returns>
         public AffidavitSaveRequest ParseWWTVFile(string filePath)
         {
-            if (!File.Exists(filePath))
-            {
-                throw new Exception("File does not exist.");
-            }
-
-            if (!Path.GetExtension(filePath).Equals(".txt"))
-            {
-                throw new Exception("Invalid file extension.");
-            }
-
-            AffidavitSaveRequest affidavitFile;
+            AffidavitSaveRequest affidavitSaveRequest = new AffidavitSaveRequest();
+            affidavitSaveRequest.FileName = Path.GetFileName(filePath);
             try
             {
-                affidavitFile = _MapWWTVFileToAffidavitFile(filePath);
-                if (affidavitFile.Details.Count == 0)
+                if (!File.Exists(filePath))
                 {
-                    return null;
+                    throw new Exception("File does not exist.");
+                }
+
+                if (!Path.GetExtension(filePath).Equals(".txt"))
+                {
+                    throw new Exception("Invalid file extension.");
+                }
+
+                _MapWWTVFileToAffidavitFile(affidavitSaveRequest,filePath);
+                if (!string.IsNullOrEmpty(_errorMessage))
+                {
+                    return affidavitSaveRequest;
                 }
             }
             catch (Exception e)
             {
                 _errorMessage = "Could not process file.\n  " + e.ToString();
-                return null;
+
+                return affidavitSaveRequest;
             }
 
-            affidavitFile.FileHash = HashGenerator.ComputeHash(File.ReadAllBytes(filePath));
-            affidavitFile.FileName = Path.GetFileName(filePath);
+            affidavitSaveRequest.FileHash = HashGenerator.ComputeHash(File.ReadAllBytes(filePath));
 
-            return affidavitFile;
+            return affidavitSaveRequest;
         }
 
-        private string _CreateInvalidFileEmailBody(string errorMessage,
-            string filePath)
+        private string _CreateInvalidFileEmailBody(string errorMessage,string filePath)
         {
             var emailBody = new StringBuilder();
 
-            emailBody.AppendFormat("File {0} failed validation for WWTV upload\n\n{1}", Path.GetFileName(filePath), errorMessage);
+            emailBody.AppendFormat("File {0} could not be properly processed.  Including technical information to help figure out the issue.", Path.GetFileName(filePath));
 
             emailBody.AppendFormat("\n\nFile located in {0}\n", filePath);
+
+            emailBody.AppendFormat("\nTechnical Information:\n\n{0}", errorMessage);
 
             return emailBody.ToString();
         }
@@ -196,13 +258,10 @@ namespace Services.Broadcast.ApplicationServices
             response.Close();
         }
 
-        private AffidavitSaveRequest _MapWWTVFileToAffidavitFile(string filePath)
+        private AffidavitSaveRequest _MapWWTVFileToAffidavitFile(AffidavitSaveRequest affidavitSaveRequest,string filePath)
         {
-            AffidavitSaveRequest file = new AffidavitSaveRequest
-            {
-                Source = (int) AffidaviteFileSource.Strata,
-                Details = new List<AffidavitSaveRequestDetail>()
-            };
+            affidavitSaveRequest.Source = (int) AffidaviteFileSource.Strata;
+            affidavitSaveRequest.Details = new List<AffidavitSaveRequestDetail>();
 
             WhosWatchingTVPostProcessingFile jsonFile;
             try
@@ -214,12 +273,12 @@ namespace Services.Broadcast.ApplicationServices
                 _errorMessage =
                     "File is in an invalid format.  It cannot be read in its current state; must be a valid JSON file." +
                     "\r\n" + e.ToString();
-                return file;
+                return affidavitSaveRequest;
             }
 
-            for (var lineNumber = 0; lineNumber < jsonFile.Details.Count; lineNumber++)
+            for (var recordNumber = 0; recordNumber < jsonFile.Details.Count; recordNumber++)
             {
-                var jsonDetail = jsonFile.Details[lineNumber];
+                var jsonDetail = jsonFile.Details[recordNumber];
 
                 var affidavitSaveRequestDetail = new AffidavitSaveRequestDetail()
                 {
@@ -257,10 +316,10 @@ namespace Services.Broadcast.ApplicationServices
 
                 }
 
-                file.Details.Add(affidavitSaveRequestDetail);
+                affidavitSaveRequest.Details.Add(affidavitSaveRequestDetail);
             }
 
-            return file;
+            return affidavitSaveRequest;
         }
 
         private void _DownloadFileFromWWTVFtp(string fileName, string filePath)
